+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added `ScheduleTime->interpolate()` to execute closure. Action after schadule execute.
+- Added DI container `Container::class`.
+- Added core framework `Appcliaction::class`, witch load DI and application configuration.
+- Added abstract `Karnel::class` and `ServiceProvider::class`.
+- Added global function Application.
 
 ## [0.3.0] - 2022-04-06
 ### Added